fixed interplay between top and bottom nav link coloring. Refactored Header component a bit more
